cleaned up int types

+++ src/Application.cpp
@@ -1,7 +1,7 @@
 #include "Application.h"
 #include <imgui.h>
 
-AppLoader::AppLoader(size_t width, size_t height, Application*& app) : Application(width, height), m_currentApp(app) { m_name = "Menu"; }
+AppLoader::AppLoader(unsigned int width, unsigned int height, Application*& app) : Application(width, height), m_currentApp(app) { m_name = "Menu"; }
 
 void AppLoader::renderImGui() {
 	for (auto& app : m_apps)
